refactor: Extract runner.rs
