+++ src/lib.rs
@@ -8,3 +8,4 @@ pub mod corelib;
 pub mod type_checking;
 pub mod error;
 pub mod names;
+pub mod runner;
